+++ src/main.rs
@@ -2,11 +2,12 @@ pub mod database;
 pub mod environment;
 pub mod routes;
 pub mod utilities;
+pub mod authenticate;
 
 #[async_std::main]
 async fn main() {
     database::connect().await;
     warp::serve(routes::routes())
-        .run(([0, 0, 0, 0], 9000))
+        .run(([0, 0, 0, 0], 29502))
         .await;
 }
